Bugfix: Shrinking overlaps
Fixed a bug where moving a sprite over its own trail caused the sprite
to shrink. The flaw in the logic was that every movement increased size,
so after X movements size would be size + X, and would need to be cut
down to max size. But not every move increases size.

`cargo fmt` done as well