Bugfix: Card duplication glitch
Details: previous to this, if you load a card and then attempt to load a
different card that you are out of, it will still decrease the count
used of the card you loaded as if it had been unloaded. This allows you
to load any number of access points with a single card, even if you only
have one copy of it.